Upgrade ScalaJS

+++ modules/core/js/src/main/scala/cron4s/lib/js/JsDateInstance.scala
@@ -75,7 +75,7 @@ private[js] final class JsDateInstance extends IsDateTime[Date] {
         dayOfWeek
     }
 
-    Right(value)
+    Right(value.toInt)
   }
 
   override def set[F <: CronField](
